requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1751547525079/work
altair @ file:///C:/b/abs_0b1avjyzoc/croot/altair_1743016738596/work
astunparse @ file:///home/conda/feedstock_root/build_artifacts/astunparse_1736248061654/work
attrs @ file:///C:/miniconda3/conda-bld/attrs_1761211086107/work
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
Bottleneck @ file:///C:/b/abs_816hr2khp1/croot/bottleneck_1731058648110/work
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools @ file:///C:/b/abs_fcvwpc5vh9/croot/cachetools_1738224687938/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1759648874697/work/certifi
cffi @ file:///D:/bld/cffi_1761202875932/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
click @ file:///D:/bld/click_1758270401364/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
contourpy @ file:///D:/bld/contourpy_1744743084386/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1755674618058/work
flatbuffers @ file:///home/conda/feedstock_root/build_artifacts/python-flatbuffers_1733838640534/work
fonttools @ file:///D:/bld/fonttools_1759187070181/work
gast @ file:///home/conda/feedstock_root/build_artifacts/gast_1733299680490/work
gitdb @ file:///C:/b/abs_1bkmriimss/croot/gitdb_1753362133190/work
GitPython @ file:///C:/miniconda3/conda-bld/gitpython_1761291330593/work
google-pasta @ file:///home/conda/feedstock_root/build_artifacts/google-pasta_1733852499742/work
greenlet==3.2.4
grpcio @ file:///D:/bld/grpc-split_1741419195867/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
h5py @ file:///D:/bld/h5py_1760616392060/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
ipython-genutils==0.2.0
ipython-sql==0.5.0
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1733308265247/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jsonschema @ file:///C:/b/abs_4d8dstds3t/croot/jsonschema_1753206424024/work
jsonschema-specifications @ file:///C:/b/abs_0brvm6vryw/croot/jsonschema-specifications_1699032417323/work
keras @ file:///home/conda/feedstock_root/build_artifacts/keras_1755182020562/work
kiwisolver @ file:///D:/bld/bld/rattler-build_kiwisolver_1756467536/work
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1757093412127/work
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1754951200865/work
MarkupSafe @ file:///D:/bld/markupsafe_1759055175616/work
matplotlib==3.10.7
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
ml-dtypes @ file:///D:/bld/ml_dtypes_1756742121008/work
munkres==1.1.4
namex @ file:///home/conda/feedstock_root/build_artifacts/namex_1748346368839/work
narwhals @ file:///C:/miniconda3/conda-bld/narwhals_1760003512045/work
numexpr @ file:///C:/b/abs_4e0h6i5ip2/croot/numexpr_1752521729801/work
numpy @ file:///C:/b/abs_0123vcxhf8/croot/numpy_and_numpy_base_1725470331966/work/dist/numpy-2.0.1-cp310-cp310-win_amd64.whl#sha256=e824ec3b279f4c5207736c83ac2f2dc79b489e08501a4ea9e3fc178c45db289b
opt_einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1733687912731/work
optree @ file:///D:/bld/optree_1756811994305/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///C:/miniconda3/conda-bld/pandas_1759822263292/work/dist/pandas-2.3.3-cp310-cp310-win_amd64.whl#sha256=1334d37cdb28f0f10d322297a276c0b56e115ad32f5bf3f0e575560f61a5ea74
pillow @ file:///C:/b/abs_b06nvnir5m/croot/pillow_1752524543576/work
prettytable==3.16.0
protobuf @ file:///D:/bld/protobuf_1741124611627/work/bazel-bin/python/dist/protobuf-5.29.3-cp310-abi3-win_amd64.whl#sha256=848a43f098da98f421273bdedee3eec14d139c91d244549d1086d16f0dafebd9
pyarrow @ file:///C:/miniconda3/conda-bld/pyarrow_1759833600583/work/python
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyparsing_1758436411/work
PySide6==6.7.3
PySocks @ file:///D:/bld/pysocks_1733217287171/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytz @ file:///C:/b/abs_f8wdzeix0n/croot/pytz_1752135878094/work
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
referencing @ file:///C:/b/abs_09f4hj6adf/croot/referencing_1699012097448/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1755614211359/work
rich @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich_1760026188/work/dist
rpds-py @ file:///C:/b/abs_0c6z5kcdb6/croot/rpds-py_1736545465023/work
shiboken6==6.7.3
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap @ file:///C:/b/abs_85z04amxrn/croot/smmap_1752748960667/work
SQLAlchemy==2.0.43
sqlparse==0.5.3
streamlit @ file:///C:/miniconda3/conda-bld/streamlit-split_1758803710627/work/streamlit
tenacity @ file:///C:/miniconda3/conda-bld/tenacity_1760375743144/work
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1740522445235/work/tensorboard-2.19.0-py3-none-any.whl#sha256=5e71b98663a641a7ce8a6e70b0be8e1a4c0c45d48760b076383ac4755c35b9a0
tensorboard_data_server @ file:///D:/bld/tensorboard-data-server_1759413103393/work/tensorboard_data_server-0.7.0-py3-none-any.whl#sha256=efeb0cc783f516ea97f2110a090757fea3a47df4a628ef76613e5d791b8d7e70
tensorflow_cpu==2.19.1
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1746039537955/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tornado @ file:///D:/bld/tornado_1756854837538/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
unicodedata2 @ file:///D:/bld/unicodedata2_1756494469536/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
watchdog @ file:///C:/miniconda3/conda-bld/watchdog_1759775793959/work
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1733160440960/work
win_inet_pton @ file:///D:/bld/win_inet_pton_1733130564612/work
wrapt @ file:///D:/bld/wrapt_1760964163245/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zstandard==0.25.0
